+++ R/buildComponent.R
@@ -225,8 +225,7 @@ ped2com <- function(ped, component,
   if (config$component %in% c("common nuclear")) {
     Matrix::diag(isPar) <- 1
 
-    if(!is.null(config$keep_ids)) {
-
+    if (!is.null(config$keep_ids)) {
       isPar <- .subsetKeepIds(
         component = isPar,
         keep_ids = keep_ids,
@@ -236,13 +235,12 @@ ped2com <- function(ped, component,
         verbose_message = "Subsetting adjacency matrix to %d target individuals\n"
       ) # also need to drop columns here because the adjacency matrix is used in the path tracing and we want to make sure the paths are correct for the target individuals.  We will keep all columns for the path tracing but subset to the target rows so that the relatedness values are correct for the target individuals.
 
-      if(length(rownames(isPar)>1)) {
+      if (length(rownames(isPar)) > 1) {
         isPar <- isPar[, rownames(isPar), drop = FALSE]
-      } #else {
-  #      isPar <- isPar[rownames(isPar)]
-  #    }
-    #  isPar <- isPar[, rownames(isPar)] #
-
+      } # else {
+      #      isPar <- isPar[rownames(isPar)]
+      #    }
+      #  isPar <- isPar[, rownames(isPar)] #
     }
     if (config$sparse == FALSE) {
       isPar <- as.matrix(isPar)